\name{practice-package}
\alias{practice-package}
\alias{practice}
\docType{package}
\title{
\packageTitle{practice}
}
\description{
\packageDescription{practice}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{practice}
\packageIndices{practice}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{practice}

Maintainer: \packageMaintainer{practice}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
